+++ crates/tauri-store/src/prelude.rs
@@ -1,4 +1,7 @@
 pub use crate::{
-  BoxResult, Error, Handle, Json, OnLoadFn, Result, SaveStrategy, Store, StoreCollection, StoreId,
-  StoreOptions, StoreState, WatcherId,
+  BoxResult, Error, Handle, Json, JsonMarshaler, Marshaler, MarshalingError, OnLoadFn, Result,
+  SaveStrategy, Store, StoreCollection, StoreId, StoreOptions, StoreState, WatcherId,
 };
+
+#[cfg(feature = "marshaler-toml")]
+pub use crate::TomlMarshaler;